sync: Import upstream OpenBSD 4.9
Import OpenBSD 4.9 and make the current L4 port reflect the changes.

Important changes affecting the L4 port include:

 - ci_cpuid converts from 32 bits to uint64_t
 - netisrs are handled as generic softintrs at IPL_SOFTNET
 - evcount_intr for evcount_attach() is gone for good
 - PCIe support (which is not ported yet)

For a more complete list of changes, visit:

  http://www.openbsd.org/plus49.html

+++ sys/arch/i386/l4/pmap.c
@@ -3132,13 +3132,13 @@ pmap_tlb_shootpage(struct pmap *pm, vaddr_t va)
 	struct cpu_info *ci, *self = curcpu();
 	CPU_INFO_ITERATOR cii;
 	int wait = 0;
-	int mask = 0;
+	uint64_t mask = 0;
 
 	CPU_INFO_FOREACH(cii, ci) {
 		if (ci == self || !pmap_is_active(pm, ci) ||
 		    !(ci->ci_flags & CPUF_RUNNING))
 			continue;
-		mask |= 1 << ci->ci_cpuid;
+		mask |= 1ULL << ci->ci_cpuid;
 		wait++;
 	}
 
@@ -3151,7 +3151,7 @@ pmap_tlb_shootpage(struct pmap *pm, vaddr_t va)
 		}
 		tlb_shoot_addr1 = va;
 		CPU_INFO_FOREACH(cii, ci) {
-			if ((mask & 1 << ci->ci_cpuid) == 0)
+			if ((mask & 1ULL << ci->ci_cpuid) == 0)
 				continue;
 			if (i386_fast_ipi(ci, LAPIC_IPI_INVLPG) != 0)
 				panic("pmap_tlb_shootpage: ipi failed");
@@ -3169,14 +3169,14 @@ pmap_tlb_shootrange(struct pmap *pm, vaddr_t sva, vaddr_t eva)
 	struct cpu_info *ci, *self = curcpu();
 	CPU_INFO_ITERATOR cii;
 	int wait = 0;
-	int mask = 0;
+	uint64_t mask = 0;
 	vaddr_t va;
 
 	CPU_INFO_FOREACH(cii, ci) {
 		if (ci == self || !pmap_is_active(pm, ci) ||
 		    !(ci->ci_flags & CPUF_RUNNING))
 			continue;
-		mask |= 1 << ci->ci_cpuid;
+		mask |= 1ULL << ci->ci_cpuid;
 		wait++;
 	}
 
@@ -3190,7 +3190,7 @@ pmap_tlb_shootrange(struct pmap *pm, vaddr_t sva, vaddr_t eva)
 		tlb_shoot_addr1 = sva;
 		tlb_shoot_addr2 = eva;
 		CPU_INFO_FOREACH(cii, ci) {
-			if ((mask & 1 << ci->ci_cpuid) == 0)
+			if ((mask & 1ULL << ci->ci_cpuid) == 0)
 				continue;
 			if (i386_fast_ipi(ci, LAPIC_IPI_INVLRANGE) != 0)
 				panic("pmap_tlb_shootrange: ipi failed");
@@ -3209,12 +3209,12 @@ pmap_tlb_shoottlb(void)
 	struct cpu_info *ci, *self = curcpu();
 	CPU_INFO_ITERATOR cii;
 	int wait = 0;
-	int mask = 0;
+	uint64_t mask = 0;
 
 	CPU_INFO_FOREACH(cii, ci) {
 		if (ci == self || !(ci->ci_flags & CPUF_RUNNING))
 			continue;
-		mask |= 1 << ci->ci_cpuid;
+		mask |= 1ULL << ci->ci_cpuid;
 		wait++;
 	}
 
@@ -3227,7 +3227,7 @@ pmap_tlb_shoottlb(void)
 		}
 
 		CPU_INFO_FOREACH(cii, ci) {
-			if ((mask & 1 << ci->ci_cpuid) == 0)
+			if ((mask & 1ULL << ci->ci_cpuid) == 0)
 				continue;
 			if (i386_fast_ipi(ci, LAPIC_IPI_INVLTLB) != 0)
 				panic("pmap_tlb_shoottlb: ipi failed");
@@ -3244,13 +3244,13 @@ pmap_tlb_droppmap(struct pmap *pm)
 	struct cpu_info *ci, *self = curcpu();
 	CPU_INFO_ITERATOR cii;
 	int wait = 0;
-	int mask = 0;
+	uint64_t mask = 0;
 
 	CPU_INFO_FOREACH(cii, ci) {
 		if (ci == self || !(ci->ci_flags & CPUF_RUNNING) ||
 		    ci->ci_curpmap != pm)
 			continue;
-		mask |= 1 << ci->ci_cpuid;
+		mask |= 1ULL << ci->ci_cpuid;
 		wait++;
 	}
 
@@ -3263,7 +3263,7 @@ pmap_tlb_droppmap(struct pmap *pm)
 		}
 
 		CPU_INFO_FOREACH(cii, ci) {
-			if ((mask & 1 << ci->ci_cpuid) == 0)
+			if ((mask & 1ULL << ci->ci_cpuid) == 0)
 				continue;
 			if (i386_fast_ipi(ci, LAPIC_IPI_RELOADCR3) != 0)
 				panic("pmap_tlb_droppmap: ipi failed");
